refactor: 디랙토리 정리
