+++ src/main/java/domain/Janggi.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.position.Position;
+import domain.position.Route;
 import java.util.ArrayList;
 import java.util.List;
 import domain.unit.Unit;
