+++ src/main/java/domain/Team.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.position.Position;
+
 public enum Team {
     HAN,
     CHO,
