+++ src/test/java/domain/unit/CarUnitRuleTest.java
@@ -1,7 +1,7 @@
 package domain.unit;
 
-import domain.Position;
-import domain.Route;
+import domain.position.Position;
+import domain.position.Route;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
