+++ src/test/java/domain/unit/PositionTest.java
@@ -1,6 +1,6 @@
 package domain.unit;
 
-import domain.Position;
+import domain.position.Position;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
